requirements.txt
llama-index
google-generativeai
llama-index-llms-gemini
pypdf
python-dotenv
IPython
llama-index-embeddings-gemini
llama-index-vector-stores-postgres
streamlit
streamlit-authenticator

python-docx
PyMuPDF
bcrypt
bleach